URLs with escaped characters are being removed by XSS because of decoding during link building. * Added escaping during link sanitization * Preserves Adobe Campaign expressions
